Master -> main

+++ PersonalCollectionManager.Infrastructure/Services/ItemService.cs
@@ -2,6 +2,7 @@
 
 using AutoMapper;
 using Microsoft.Extensions.Logging;
+using PersonalCollectionManager.Application.DTOs;
 using PersonalCollectionManager.Application.DTOs.RequestDtos;
 using PersonalCollectionManager.Application.DTOs.ResponseDtos;
 using PersonalCollectionManager.Application.Interfaces.IRepository;
@@ -14,27 +15,30 @@ namespace PersonalCollectionManager.Infrastructure.Services
     public class ItemService : IItemService
     {
         private readonly IItemRepository _itemRepository;
+        private readonly IItemTagRepository _itemTagRepository;
         private readonly IMapper _mapper;
         private readonly ILogger<ItemService> _logger;
-        public ItemService(IItemRepository itemRepository,IMapper mapper, ILogger<ItemService> logger)
+        public ItemService(IItemRepository itemRepository,IItemTagRepository itemTagRepository,IMapper mapper, ILogger<ItemService> logger)
         {
             _itemRepository = itemRepository;
+            _itemTagRepository = itemTagRepository;
             _mapper = mapper;
             _logger = logger;
         }
 
-        public async Task<OperationResult> AddItemAsync(ItemRequestDto item)
+        public async Task<ItemDto> AddItemAsync(ItemRequestDto item)
         {
             try
             {
-                var user = _mapper.Map<Item>(item);
-                await _itemRepository.AddAsync(user);
-                return new OperationResult(true, "Item added successfully");
+                var items = _mapper.Map<Item>(item);
+                items.DateAdded = DateTime.Now;
+                var result = await _itemRepository.AddAsync(items);
+                return _mapper.Map<ItemDto>(result);
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error adding item");
-                return new OperationResult(false, "Error adding item");
+                return null;
             }
         }
 
@@ -43,7 +47,11 @@ namespace PersonalCollectionManager.Infrastructure.Services
             try
             {
                 var user = await _itemRepository.GetByIdAsync(id);
-                _itemRepository.Remove(user);
+                if (user == null)
+                {
+                    return new OperationResult(false, "Item not found");
+                }
+                await _itemRepository.Remove(user);
                 return new OperationResult(true, "Item deleted successfully");
             }
             catch (Exception ex)
@@ -53,12 +61,26 @@ namespace PersonalCollectionManager.Infrastructure.Services
             }
         }
 
-        public async Task<IEnumerable<ItemDTO>> GetAllItemsAsync()
+        public async Task<IEnumerable<ItemDto>> GetAllItemByCollectionIdAsync(Guid id)
+        {
+            try
+            {
+                var user = await _itemRepository.GetItemsByCollectionIdAsync(id);
+                return _mapper.Map<IEnumerable<ItemDto>>(user);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "Error getting all items by collection id");
+                throw;
+            }
+        }
+
+        public async Task<IEnumerable<ItemDto>> GetAllItemsAsync()
         {
             try
             {
                 var user = await _itemRepository.GetAllItemsAsync();
-                return _mapper.Map<IEnumerable<ItemDTO>>(user);
+                return _mapper.Map<IEnumerable<ItemDto>>(user);
             }
             catch (Exception ex)
             {
@@ -67,12 +89,12 @@ namespace PersonalCollectionManager.Infrastructure.Services
             }
         }
 
-        public async Task<ItemDTO> GetItemByIdAsync(Guid id)
+        public async Task<ItemDto> GetItemByIdAsync(Guid id)
         {
             try
             {
                 var user = await _itemRepository.GetByIdAsync(id);
-                return _mapper.Map<ItemDTO>(user);
+                return _mapper.Map<ItemDto>(user);
             }
             catch (Exception ex)
             {
@@ -81,18 +103,47 @@ namespace PersonalCollectionManager.Infrastructure.Services
             }
         }
 
-        public async Task<OperationResult> UpdateItemAsync(ItemRequestDto item)
+        public async Task<IEnumerable<ItemDto>> GetItemsByTagAsync(string tagName)
+        {
+            try
+            {
+                var items = await _itemRepository.GetItemsByTagAsync(tagName);
+                return _mapper.Map<IEnumerable<ItemDto>>(items);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "Error getting items by tag");
+                throw;
+            }
+        }
+
+        public async Task<IEnumerable<ItemDto>> GetRecentItemsAsync()
+        {
+            try
+            {
+                var item =await _itemRepository.GetRecentItemsAsync();
+
+                return _mapper.Map<IEnumerable<ItemDto>>(item);
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "Error getting recent items");
+                throw;
+            }
+        }
+
+        public async Task<ItemDto> UpdateItemAsync(ItemDto item)
         {
             try
             {
                 var user = _mapper.Map<Item>(item);
-                _itemRepository.Update(user);
-                return new OperationResult(true, "Item updated successfully");
+                var result = await _itemRepository.Update(user);
+                return _mapper.Map<ItemDto>(result);
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error Updating Item");
-                throw;
+                return null;
             }
         }
